Use the hash of the content for the unique constraints.
Postgres doesn't like unique constraints involving very long text fields. Instead of requiring that title and description (which can be very long) are unique for an embedding, we hash the entire content of the submission and store the hash with it and just require that the hash is unique. We also include the abstract. This is almost equivalent to the previous approach, but works better with Postgres unique constraints.

On migration, the hash is computed for every existing embedding.